requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.10.0
attrs==25.4.0
baron==0.10.1
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.8
django-cache-url==3.4.5
django-cors-headers==4.9.0
django-environ==0.12.0
django-environs==0.1.1
django-filter==25.2
django-heroku==0.3.1
django-model-utils==5.0.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
environs==14.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
marshmallow==4.1.0
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redbaron==0.9.2
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
rply==0.7.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
